Added JSDoc comments
